+++ R/Immo_class.R
@@ -9,18 +9,16 @@
 #----------------------------------------------------------------------------------------------------------------------------------------------------
 #           Declarateur
 #----------------------------------------------------------------------------------------------------------------------------------------------------
-##' La classe Immo
-##'
-##' Classe pour les actifs de type Immo
+##' Classe pour les actifs de type immobilier.
 ##'
 ##' @name Immo
-##' @slot ptf_immo est un dataframe, chaque ligne represente un actif immobilier du portefeuille d'Immobilier.
+##' @slot ptf_immo est un dataframe, chaque ligne represente un actif immobilier du portefeuille d'immobilier.
 ##' @docType class
 ##' @section Lien a creer
 ##' @author Prim'Act
 ##' @seealso Mettre le lien vers les methodes de la classe
-##' @keywords classes
 ##' @export
+##' @seealso Les operations d'achat vente immo  \code{\link{buy_immo}} et \code{\link{sell_immo}}.
 setClass(
   Class = "Immo",
   representation = representation(
@@ -40,16 +38,16 @@ setValidity ("Immo",
 
                # Verification du type des colonnes
                if (!is.integer(object@ptf_immo[,1]))  {retval <- c(retval, "[Immo] : num_mp n'est pas entier/n")}
-               if (!is.double(object@ptf_immo[,2]))   {retval <- c(retval, "[Immo] : val_marche n'est pas reel/n")}
-               if (!is.double(object@ptf_immo[,3]))   {retval <- c(retval, "[Immo] : val_nc n'est pas reel/n")}
-               if (!is.double(object@ptf_immo[,4]))   {retval <- c(retval, "[Immo] : val_achat n'est pas reel/n")}
+               if (!is.numeric(object@ptf_immo[,2]))   {retval <- c(retval, "[Immo] : val_marche n'est pas reel/n")}
+               if (!is.numeric(object@ptf_immo[,3]))   {retval <- c(retval, "[Immo] : val_nc n'est pas reel/n")}
+               if (!is.numeric(object@ptf_immo[,4]))   {retval <- c(retval, "[Immo] : val_achat n'est pas reel/n")}
                if (!is.logical(object@ptf_immo[,5]))  {retval <- c(retval, "[Immo] : presence n'est pas logical/n")}
                if (!is.logical(object@ptf_immo[,6]))  {retval <- c(retval, "[Immo] : cessible n'est pas logical/n")}
-               if (!is.double(object@ptf_immo[,7]))   {retval <- c(retval, "[Immo] : nb_unit n'est pas reel/n")}
-               if (!is.double(object@ptf_immo[,8]))   {retval <- c(retval, "[Immo] : dur_det n'est pas reel/n")}
-               if (!is.double(object@ptf_immo[,9]))   {retval <- c(retval, "[Immo] : pdd n'est pas reel/n")}
+               if (!is.numeric(object@ptf_immo[,7]))   {retval <- c(retval, "[Immo] : nb_unit n'est pas reel/n")}
+               if (!is.numeric(object@ptf_immo[,8]))   {retval <- c(retval, "[Immo] : dur_det n'est pas reel/n")}
+               if (!is.numeric(object@ptf_immo[,9]))   {retval <- c(retval, "[Immo] : pdd n'est pas reel/n")}
                if (!is.integer(object@ptf_immo[,10])) {retval <- c(retval, "[Immo] : num_index n'est pas integer/n")}
-               if (!is.double(object@ptf_immo[,11]))  {retval <- c(retval, "[Immo] : loyer n'est pas reel/n")}
+               if (!is.numeric(object@ptf_immo[,11]))  {retval <- c(retval, "[Immo] : loyer n'est pas reel/n")}
                if (!is.logical(object@ptf_immo[,12])) {retval <- c(retval, "[Immo] : ind_invest n'est pas logical/n")}
 
                # Verification du nom des colonnes
@@ -77,9 +75,9 @@ setMethod(
                                            "pdd","num_index","loyer","ind_invest"))!=0)
       {stop("[Immo] : Nombre ou nommage des colonnes du dataframe incorrect")}
       else if(
-        !is.integer(ptf[,"num_mp"])   | !is.double(ptf[,"val_marche"]) | !is.double(ptf[,"val_nc"])  | !is.double(ptf[,"val_achat"]) |
-        !is.logical(ptf[,"presence"]) | !is.logical(ptf[,"cessible"])  | !is.double(ptf[,"nb_unit"]) | !is.double(ptf[,"dur_det"])   |
-        !is.double(ptf[,"pdd"])       | !is.integer(ptf[,"num_index"]) | !is.double(ptf[,"loyer"])     | !is.logical(ptf[,"ind_invest"]))
+        !is.integer(ptf[,"num_mp"])   | !is.numeric(ptf[,"val_marche"]) | !is.numeric(ptf[,"val_nc"])  | !is.numeric(ptf[,"val_achat"]) |
+        !is.logical(ptf[,"presence"]) | !is.logical(ptf[,"cessible"])  | !is.numeric(ptf[,"nb_unit"]) | !is.numeric(ptf[,"dur_det"])   |
+        !is.numeric(ptf[,"pdd"])       | !is.integer(ptf[,"num_index"]) | !is.numeric(ptf[,"loyer"])     | !is.logical(ptf[,"ind_invest"]))
       {stop("[Immo] : Typage incorrect des colonnes du dataframe")}
       else
       {.Object@ptf_immo <- ptf
@@ -87,7 +85,7 @@ setMethod(
     }
     #Traitement du cas vide
     else
-    {.Object@ptf_immo  <- data.frame(integer(),double(),double(),double(),logical(),logical(),double(),double(),double(),integer(),double(),logical())
+    {.Object@ptf_immo  <- data.frame(integer(),numeric(),numeric(),numeric(),logical(),logical(),numeric(),numeric(),numeric(),integer(),numeric(),logical())
     colnames(.Object@ptf_immo) <- c("num_mp","val_marche","val_nc","val_achat","presence","cessible","nb_unit","dur_det","pdd","num_index","loyer","ind_invest")
     }
     return(.Object)
