+++ R/Oblig_class.r
@@ -6,23 +6,20 @@
 #--------------------------------------------------------------------------------------------------------------------
 
 
-
 #----------------------------------------------------------------------------------------------------------------------------------------------------
 #           Declarateur
 #----------------------------------------------------------------------------------------------------------------------------------------------------
-##' La classe Oblig
-##'
-##' Classe pour les actifs de type Obligation
+##' Classe pour les actifs de type obligation.
 ##'
 ##' @name Oblig
-##' @slot ptf_oblig est un dataframe, chaque ligne represente un actif obligataire du portefeuille d'obligation.
+##' @slot ptf_oblig est un dataframe, chaque ligne represente un actif obligation du portefeuille d'obligation.
 ##' @docType class
 ##' @section Lien a creer
 ##' @author Prim'Act
 ##' @seealso Mettre le lien vers les methodes de la classe
-##' @keywords classes
 ##' @export
-#removeClass("Oblig")
+##' @seealso Les operations d'achat vente obligations  \code{\link{buy_oblig}} et \code{\link{sell_oblig}}.
+
 setClass(
   Class = "Oblig",
   representation = representation(
@@ -42,23 +39,23 @@ setValidity(Class = "Oblig",
 
               # Verification du type des colonnes
               if (!is.integer(object@ptf_oblig[,1]))  {retval <- c(retval, "[Oblig] : num_mp n'est pas entier/n")}
-              if (!is.double(object@ptf_oblig[,2]))   {retval <- c(retval, "[Oblig] : val_marche n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,3]))   {retval <- c(retval, "[Oblig] : val_nc n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,4]))   {retval <- c(retval, "[Oblig] : val_achat n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,2]))   {retval <- c(retval, "[Oblig] : val_marche n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,3]))   {retval <- c(retval, "[Oblig] : val_nc n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,4]))   {retval <- c(retval, "[Oblig] : val_achat n'est pas reel/n")}
               if (!is.logical(object@ptf_oblig[,5]))  {retval <- c(retval, "[Oblig] : presence n'est pas reel/n")}
               if (!is.logical(object@ptf_oblig[,6]))  {retval <- c(retval, "[Oblig] : cessible n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,7]))   {retval <- c(retval, "[Oblig] : nb_unit n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,8]))   {retval <- c(retval, "[Oblig] : dur_det n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,9]))   {retval <- c(retval, "[Oblig] : nominal n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,10]))  {retval <- c(retval, "[Oblig] : Taux coupon n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,11]))  {retval <- c(retval, "[Oblig] : Parite n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,12]))  {retval <- c(retval, "[Oblig] : Maturite residuelle n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,7]))   {retval <- c(retval, "[Oblig] : nb_unit n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,8]))   {retval <- c(retval, "[Oblig] : dur_det n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,9]))   {retval <- c(retval, "[Oblig] : nominal n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,10]))  {retval <- c(retval, "[Oblig] : Taux coupon n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,11]))  {retval <- c(retval, "[Oblig] : Parite n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,12]))  {retval <- c(retval, "[Oblig] : Maturite residuelle n'est pas reel/n")}
               if (!is.factor(object@ptf_oblig[,13]))  {retval <- c(retval, "[Oblig] : Type n'est pas factor/n")}
               if (!is.integer(object@ptf_oblig[,14])) {retval <- c(retval, "[Oblig] : Rating n'est pas integer/n")}
-              if (!is.double(object@ptf_oblig[,15]))  {retval <- c(retval, "[Oblig] : Duration n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,16]))  {retval <- c(retval, "[Oblig] : Zspread n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,17]))  {retval <- c(retval, "[Oblig] : Coupon couru n'est pas reel/n")}
-              if (!is.double(object@ptf_oblig[,18]))  {retval <- c(retval, "[Oblig] : Surcote/Decote n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,15]))  {retval <- c(retval, "[Oblig] : Duration n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,16]))  {retval <- c(retval, "[Oblig] : Zspread n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,17]))  {retval <- c(retval, "[Oblig] : Coupon couru n'est pas reel/n")}
+              if (!is.numeric(object@ptf_oblig[,18]))  {retval <- c(retval, "[Oblig] : Surcote/Decote n'est pas reel/n")}
 
               # Verification du nom des colonnes
               if(sum(colnames(object@ptf_oblig)==c("num_mp","val_marche","val_nc","val_achat",
@@ -87,18 +84,18 @@ setMethod(
                                                    "rating","duration","zspread","cc","sd")) != 0)
       {stop("[Oblig] : Nombre ou nommage des colonnes du dataframe incorrect")}
       else if(
-        !is.integer(ptf[,"num_mp"])   | !is.double(ptf[,"val_marche"]) | !is.double(ptf[,"val_nc"])  | !is.double(ptf[,"val_achat"]) |
-        !is.logical(ptf[,"presence"]) | !is.logical(ptf[,"cessible"])  | !is.double(ptf[,"nb_unit"]) | !is.double(ptf[,"dur_det"])   |
-        !is.double(ptf[,"nominal"])   | !is.double(ptf[,"tx_coupon"])  | !is.double(ptf[,"par"])     | !is.double(ptf[,"mat_res"])   |
-        !is.factor(ptf[,"type"])      | !is.integer(ptf[,"rating"])    | !is.double(ptf[,"duration"])| !is.double(ptf[,"zspread"])   |
-        !is.double(ptf[,"cc"])        | !is.double(ptf[,"sd"]))  {stop("[Oblig] : Typage incorrect des colonnes du dataframe")}
+        !is.integer(ptf[,"num_mp"])   | !is.numeric(ptf[,"val_marche"]) | !is.numeric(ptf[,"val_nc"])  | !is.numeric(ptf[,"val_achat"]) |
+        !is.logical(ptf[,"presence"]) | !is.logical(ptf[,"cessible"])  | !is.numeric(ptf[,"nb_unit"]) | !is.numeric(ptf[,"dur_det"])   |
+        !is.numeric(ptf[,"nominal"])   | !is.numeric(ptf[,"tx_coupon"])  | !is.numeric(ptf[,"par"])     | !is.numeric(ptf[,"mat_res"])   |
+        !is.factor(ptf[,"type"])      | !is.integer(ptf[,"rating"])    | !is.numeric(ptf[,"duration"])| !is.numeric(ptf[,"zspread"])   |
+        !is.numeric(ptf[,"cc"])        | !is.numeric(ptf[,"sd"]))  {stop("[Oblig] : Typage incorrect des colonnes du dataframe")}
       else
       {.Object@ptf_oblig <- ptf
       validObject(.Object)}
     }
     #Traitement du cas vide
     else
-    {.Object@ptf_oblig  <- data.frame(integer(),double(),double(),double(),logical(),logical(),double(),double(),double(),double(),double(),double(),factor(),integer(),double(),double(),double(),double())
+    {.Object@ptf_oblig  <- data.frame(integer(),numeric(),numeric(),numeric(),logical(),logical(),numeric(),numeric(),numeric(),numeric(),numeric(),numeric(),factor(),integer(),numeric(),numeric(),numeric(),numeric())
     colnames(.Object@ptf_oblig) <- c("num_mp","val_marche","val_nc","val_achat","presence","cessible","nb_unit","dur_det","nominal","tx_coupon","par","mat_res","type","rating","duration","zspread","cc","sd")
     }
     return(.Object)
